fastapi/requirements.txt
fastapi
uvicorn
python-multipart
minio
pydantic
python-dotenv
autodistill-grounded-sam
autodistill-yolov8
roboflow
boto3
supervision==0.9.0
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0